Create E10_PythonCode
This code should be run when the E10 is powered on. 
It initializes a connection into the SNAP network and a remote connection with ISCA's web server.
Then this runs silently in the background waiting for a bug catch event on a node in the SNAP network.
When an event fires, this updates appropriate files on ISCA's web server.